acpi: fix acpi power button

+++ board/recalbox/recalbox-patch-target.sh
@@ -22,6 +22,12 @@ ln -sf "/recalbox/share/cheats"                                       "${TARGET_
 # we don't want the kodi startup script
 rm -f "${TARGET_DIR}/etc/init.d/S50kodi" || exit 1
 
+# acpid requires /var/run, so, requires S03populate
+if test -e "${TARGET_DIR}/etc/init.d/S02acpid"
+then
+    mv "${TARGET_DIR}/etc/init.d/S02acpid" "${TARGET_DIR}/etc/init.d/S05acpid" || exit 1
+fi
+
 # we don't want default xorg files
 rm -f "${TARGET_DIR}/etc/X11/xorg.conf" || exit 1
